backend/requirements.txt
pandas
numpy
scikit-learn
joblib
flask
flask-cors
sqlalchemy
pymysql
python-dotenv
